requirements.txt
django==5.2.6
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
drf-spectacular==0.28.0
django-filter==25.1
django-unfold==0.66.0
unidecode==1.4.0
psycopg[binary]==3.2.10
gunicorn==23.0.0
environs==14.3.0
requests==2.32.5
pyyaml==6.0.3
celery==5.5.3
django-debug-toolbar==6.0.0
